+++ librecraft/src/gui/widget/selectableList.cpp
@@ -22,38 +22,40 @@ using namespace graphics;
 
 namespace widget {
 
-SelectableList::SelectableList(int id, int x, int y, int width, int height,
-		function<void(int)> observer, int layer)
-		: AbstractWidget(id, x, y, width, height), m_layer {layer + 1} {
+SelectableList::SelectableList(int id, int x, int y, int width, int height, function<void(int)> observer,
+        unsigned layer)
+    : AbstractWidget(id, static_cast<double>(x), static_cast<double>(y), width, height), m_layer{layer + 1}
+{
 	this->m_observer = observer;
 
 	auto &res = Resources::getInstance();
 
-	m_sprite.reset(
-			new Sprite {x, y, layer, width, height, res.getTexture(
-					config::gui_data::button)});
+	m_sprite.reset(new Sprite{static_cast<double>(x), static_cast<double>(y), layer,
+	    static_cast<double>(width), static_cast<double>(height),
+	    res.getTexture(config::gui_data::button)});
 
 }
 
 void SelectableList::addListItem(std::string item) {
 
-	m_buttons.push_back(
-			shared_ptr<ToggleButton>(
-					new ToggleButton(++idCounter, m_xCoordinate + 5,
-							m_yCoordinate + m_height
-									- (5 + 30 * (1 + m_buttons.size())),
-							m_width - 10, 30, [&](int id)
-							{
-								auto button = getButtonWithId(id);
-								if (button->isToggled()) {
-									if (m_currentlyToggled)
-									m_currentlyToggled->toggle();
+    // TODO WTF IS THIS LOL
+    // TODO Fix so that it does not look completely awful.
+	m_buttons.push_back(shared_ptr<ToggleButton>(
+        new ToggleButton(++idCounter, m_xCoordinate + 5,
+                m_yCoordinate + m_height
+                        - (5 + 30 * (1 + m_buttons.size())),
+                m_width - 10, 30, [&](int id)
+                {
+                    auto button = getButtonWithId(id);
+                    if (button->isToggled()) {
+                        if (m_currentlyToggled)
+                        m_currentlyToggled->toggle();
 
-									m_currentlyToggled = button;
-								} else {
-									m_currentlyToggled.reset();
-								}
-							}, item, m_layer)));
+                        m_currentlyToggled = button;
+                    } else {
+                        m_currentlyToggled.reset();
+                    }
+                }, item, m_layer)));
 
 }
